Reorganize the conformance tests to reduce dependencies
To avoid having j4rs in build-dependencies, we move the Java-integration
into a separate private zetasketch-java crate that we only include
in main crate's dev-dependencies. This way, j4rs is only pulled and
built when tests of zetasketch-rs are built. Regular users of the crate
won't have to pull it in.